+++ src/decompress.h
@@ -22,4 +22,4 @@
 
 void unpack_entry(const localheader &lh,
         const centralheader &ch,
-        const unsigned char *data_start, uint32_t data_size);
+        const unsigned char *data_start, uint64_t data_size);
